Refaktor veza entiteta i nekih atributa - Odradjen dosta veliki refaktor kod zavisnosti modela
- Promenjeno ime paketa iz beans u models
