+++ nvt_kts_back/src/main/java/com/example/nvt_kts_back/controllers/DriverController.java
@@ -1,6 +1,6 @@
 package com.example.nvt_kts_back.controllers;
 
-import com.example.nvt_kts_back.beans.Driver;
+import com.example.nvt_kts_back.models.Driver;
 import com.example.nvt_kts_back.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
