+++ nvt_kts_back/src/main/java/com/example/nvt_kts_back/controllers/ReviewController.java
@@ -1,6 +1,6 @@
 package com.example.nvt_kts_back.controllers;
 
-import com.example.nvt_kts_back.beans.Review;
+import com.example.nvt_kts_back.models.Review;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
