+++ nvt_kts_back/src/main/java/com/example/nvt_kts_back/controllers/RideController.java
@@ -1,6 +1,6 @@
 package com.example.nvt_kts_back.controllers;
 
-import com.example.nvt_kts_back.beans.Ride;
+import com.example.nvt_kts_back.models.Ride;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
