+++ nvt_kts_back/src/main/java/com/example/nvt_kts_back/models/Role.java
@@ -1,9 +1,8 @@
-package com.example.nvt_kts_back.beans;
+package com.example.nvt_kts_back.models;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import javax.persistence.*;
 
@@ -12,12 +11,10 @@ import javax.persistence.*;
 @NoArgsConstructor
 public class Role implements GrantedAuthority {
     private static final long serialVersionUID = 1L;
-
     @Id
     @Column
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     Long id;
-
     @Column(nullable = false)
     String name;
