+++ nvt_kts_back/src/main/java/com/example/nvt_kts_back/models/TempCodeHolder.java
@@ -1,4 +1,4 @@
-package com.example.nvt_kts_back.beans;
+package com.example.nvt_kts_back.models;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
