+++ nvt_kts_back/src/main/java/com/example/nvt_kts_back/models/User.java
@@ -1,4 +1,4 @@
-package com.example.nvt_kts_back.beans;
+package com.example.nvt_kts_back.models;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -36,7 +36,7 @@ public class User implements UserDetails {
     @Column
     private Boolean isBlocked;
     @ManyToOne(cascade = CascadeType.ALL)
-    @JoinColumn(name = "role_id", referencedColumnName = "id", nullable = false)
+    @JoinColumn(nullable = false)
     private Role role;
 
     public User(String email, String password, String name, String surname, String city, String phone,
