+++ nvt_kts_back/src/main/java/com/example/nvt_kts_back/repository/RegisteredUserRepository.java
@@ -1,6 +1,6 @@
 package com.example.nvt_kts_back.repository;
 
-import com.example.nvt_kts_back.beans.RegisteredUser;
+import com.example.nvt_kts_back.models.RegisteredUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
