+++ nvt_kts_back/src/main/java/com/example/nvt_kts_back/repository/ReviewRepository.java
@@ -1,6 +1,6 @@
 package com.example.nvt_kts_back.repository;
 
-import com.example.nvt_kts_back.beans.Review;
+import com.example.nvt_kts_back.models.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReviewRepository extends JpaRepository<Review,Integer> {
