+++ nvt_kts_back/src/main/java/com/example/nvt_kts_back/repository/RideRepository.java
@@ -1,6 +1,6 @@
 package com.example.nvt_kts_back.repository;
 
-import com.example.nvt_kts_back.beans.Ride;
+import com.example.nvt_kts_back.models.Ride;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RideRepository extends JpaRepository<Ride, Integer> {
