+++ nvt_kts_back/src/main/java/com/example/nvt_kts_back/repository/RoleRepository.java
@@ -1,6 +1,6 @@
 package com.example.nvt_kts_back.repository;
 
-import com.example.nvt_kts_back.beans.Role;
+import com.example.nvt_kts_back.models.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
