+++ nvt_kts_back/src/main/java/com/example/nvt_kts_back/repository/RouteRepository.java
@@ -1,6 +1,6 @@
 package com.example.nvt_kts_back.repository;
 
-import com.example.nvt_kts_back.beans.Route;
+import com.example.nvt_kts_back.models.Route;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RouteRepository extends JpaRepository<Route,Integer> {
