+++ nvt_kts_back/src/main/java/com/example/nvt_kts_back/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.example.nvt_kts_back.repository;
 
-import com.example.nvt_kts_back.beans.User;
+import com.example.nvt_kts_back.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
