+++ nvt_kts_back/src/main/java/com/example/nvt_kts_back/service/AuthService.java
@@ -7,19 +7,16 @@ import com.example.nvt_kts_back.CustomExceptions.UserDoesNotExistException;
 import com.example.nvt_kts_back.DTOs.AuthTokenDTO;
 import com.example.nvt_kts_back.DTOs.FBLoginDTO;
 import com.example.nvt_kts_back.DTOs.LoginDTO;
-import com.example.nvt_kts_back.beans.RegisteredUser;
-import com.example.nvt_kts_back.beans.Role;
-import com.example.nvt_kts_back.beans.User;
+import com.example.nvt_kts_back.models.RegisteredUser;
+import com.example.nvt_kts_back.models.Role;
+import com.example.nvt_kts_back.models.User;
 import com.example.nvt_kts_back.configurations.Settings;
 import com.example.nvt_kts_back.repository.RoleRepository;
 import com.example.nvt_kts_back.security.TokenUtils;
 import com.restfb.DefaultFacebookClient;
 import com.restfb.FacebookClient;
-import com.restfb.Parameter;
 import com.restfb.Version;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
