+++ nvt_kts_back/src/main/java/com/example/nvt_kts_back/service/TempCodeHolderService.java
@@ -2,7 +2,7 @@ package com.example.nvt_kts_back.service;
 
 import com.example.nvt_kts_back.CustomExceptions.InvalidTempCodeException;
 import com.example.nvt_kts_back.CustomExceptions.PasswordResetTempCodeDoesNotExistException;
-import com.example.nvt_kts_back.beans.TempCodeHolder;
+import com.example.nvt_kts_back.models.TempCodeHolder;
 import com.example.nvt_kts_back.repository.TempCodeHolderRepository;
 import com.example.nvt_kts_back.utils.temp_code_gen.TempCodeGenerator;
 import org.springframework.beans.factory.annotation.Autowired;
