+++ nvt_kts_back/src/main/java/com/example/nvt_kts_back/service/UserService.java
@@ -3,12 +3,11 @@ package com.example.nvt_kts_back.service;
 import com.example.nvt_kts_back.CustomExceptions.InvalidDTOAttributesValuesException;
 import com.example.nvt_kts_back.CustomExceptions.UserDoesNotExistException;
 import com.example.nvt_kts_back.DTOs.PasswordResetDTO;
-import com.example.nvt_kts_back.beans.Driver;
-import com.example.nvt_kts_back.beans.RegisteredUser;
-import com.example.nvt_kts_back.beans.User;
+import com.example.nvt_kts_back.models.Driver;
+import com.example.nvt_kts_back.models.RegisteredUser;
+import com.example.nvt_kts_back.models.User;
 import com.example.nvt_kts_back.repository.RegisteredUserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import com.example.nvt_kts_back.repository.UserRepository;
